Create NavigationBar Child component
Create navigation-bar folder and inside of it, create NavigationBar.jsx file. Import <Navbar> and sub-components from react-bootstrap and import <Link> from react-router-dom. Create NavigationBar child component and export it.  Display the login and signup links if user is logged out, and display the home and logout links, if user is logged in. Use the onLoggedOut prop to be called when logout is clicked.