Update MovieCard for fav movie btn
Import movie-card.scss. Create state variable, called isFavorite, where the initial state is "false". Set the isFavorite state variable to "true", if there is fav movie of the current user. Use ternary operator to check if the movie is one of the favourite movies, if yes, then return "red-heart" icon.